backend_for_lovable/requirements.txt
# Backend for Lovable - Dependencies
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Google Cloud Authentication & Vertex AI (updated version with agent_engines)
google-auth==2.27.0
google-cloud-aiplatform>=1.60.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0
